fix: use only node corresponding to a given set.
Before, if a mesh from an element set was retrieved, all nodes of the
model were present in the mesh. This was done for simplicity, as in this
way the references to from the elements to the nodes don't need to be
changed. This was improved with this commit by only selecting the nodes
that correspond to the elements in the set and renumbering the cell
defintition of each element accordingly.