feat: element CAX4 added, fix: key error, feat: new example
* New example: displacement visualization

Added a new folder with a example of displacement visualization. The uploaded folder also contains the ABAQUS input file and the result file (.fil).

* Added CAX4 element 

CAX4 (Axisymmetric, 4 nodes, bilinear) element was added on the element list.

* Small fix for only nodal results

When just the nodal result was requested in the ABAQUS input file, this part of the code raised a KeyError exception. The fix was to check if variable var was in the elem_output before assigning elem_ids (Tested with displacement_visualization.py in the Examples folder).

* Few typos

Few typos in the example code

* color="white" part deleted as suggested in the PR

* Revert "Added CAX4 element "

This reverts commit 85d76a069154100b53ac67f0aa82be1e7a7529c7.

* Fix: Change tracking for the last commit

Change tracking for the last commit

Added CAX4 element

* Revert "Small fix for only nodal results"

This reverts commit 7560583f8da500f840401efe02260077500a5c1d.

* Properly error exception handling.

Properly error exception handling.

For the added example "displacement_visualization.py", a KeyError exception was raised due to the non-existence of element results in the ".fil" file.

 try and except block was added.