fix: correctly consider set names
This commit fixes an issue occurring when parsing the name/label of a
node or element set. Originally, the parser was explicitly looking for
an integer value, which was then cross-referenced with the information
contained in record 1940. However, for the cases where the label string
has less or equal than 8 chars the label is written directly (not in
record 1940). This fixes the issue by first checking whether the record
is an integer or not, and handling the parsing of the name and
consecutive elements/nodes accordingly.